refractor: extract detect device function from Header, AgencyHomePage and SearchResults to useDetectDevice hook
