+++ client/src/pages/AgencyHomePage/AgencyHomePage.component.tsx
@@ -2,6 +2,7 @@ import { Box, Flex, HStack, Spacer, VStack, Text } from '@chakra-ui/react'
 import { ReactNode, useContext, useEffect, useState } from 'react'
 import { useQuery } from 'react-query'
 import { useParams } from 'react-router-dom'
+import { useDetectDevice } from '../../hooks/useDetectDevice'
 import AgencyLogo from '../../components/AgencyLogo/AgencyLogo.component'
 import CitizenRequest from '../../components/CitizenRequest/CitizenRequest.component'
 import PageTitle from '../../components/PageTitle/PageTitle.component'
@@ -40,34 +41,7 @@ const AgencyHomePage = (): JSX.Element => {
   )
 
   // Designer: fair to assume user will almost always edit on desktop
-  const device = {
-    mobile: 'mobile',
-    tablet: 'tablet',
-    desktop: 'desktop',
-  }
-
-  const [deviceType, setDeviceType] = useState(
-    window.innerWidth < 480
-      ? device.mobile
-      : window.innerWidth < 1440
-      ? device.tablet
-      : device.desktop,
-  )
-
-  const checkViewportSize = () => {
-    setDeviceType(
-      window.innerWidth < 480
-        ? device.mobile
-        : window.innerWidth < 1440
-        ? device.tablet
-        : device.desktop,
-    )
-  }
-
-  useEffect(() => {
-    window.addEventListener('resize', checkViewportSize)
-    return () => window.removeEventListener('resize', checkViewportSize)
-  }, [])
+  const deviceType = useDetectDevice()
 
   const bannerWithNeedHelpAndAgencyLogo = (
     <HStack
@@ -238,7 +212,7 @@ const AgencyHomePage = (): JSX.Element => {
         ? homePageDefaultView // no topics param key -> default homepage
         : !topicQueried
         ? homePageAllQuestionsView // topics param key + no topic queried -> all questions
-        : deviceType === device.desktop // specific topic selected
+        : deviceType === 'desktop' // specific topic selected
         ? topicPageDesktopView
         : topicPageMobileView}
       <Spacer />
